Move location of LSP storage helpers
The `lsp` namespace is bloated enough as it is, and the
store did not depend on it.
